Adding titles with helmet and fixing images
