+++ src/components/Header/SubjectsList.jsx
@@ -4,27 +4,10 @@ import styles from "./SubjectsList.module.scss";
 import { PropTypes } from "prop-types";
 import { useDispatch } from "react-redux";
 import { paginationActions } from "../../store/pagination";
+import subjectsLists from "../../utils/subjectsList";
 
 function SubjectsList(props) {
   const dispatch = useDispatch();
-
-  const subjectsLists = [
-    { name: "Fantasia", id: "fantasy" },
-    { name: "Cocina", id: "cooking" },
-    { name: "Romance", id: "romance" },
-    { name: "Amor", id: "love" },
-    { name: "Ciencia ficción", id: "science_fiction" },
-    { name: "Literatura juvenil", id: "juvenile_literature" },
-    { name: "Gatos", id: "cats" },
-    { name: "Programación", id: "programming" },
-    { name: "Suspenso", id: "thriller" },
-    { name: "Emprendimiento", id: "entrepreneurship" },
-    { name: "Humor", id: "humor" },
-    { name: "Literatura", id: "literature" },
-    { name: "Administración", id: "management" },
-    { name: "Ciencia ficción juvenil", id: "young_adult_fiction" },
-    { name: "Magia", id: "magic" },
-  ];
 
   const changeListDisplayHandler = () => {
     props.setDropdownClicked((prevState) => !prevState);
