+++ src/pages/BookDetail.module.scss
@@ -14,6 +14,7 @@
 .book-detail-img {
   display: flex;
   justify-content: center;
+  align-items: center;
   img {
     max-width: 100%;
     height: auto;
@@ -21,6 +22,39 @@
   }
   @include sm {
     width: 50%;
+  }
+}
+
+.book-detail-image-no-image {
+  background-color: lightgrey;
+  width: 60%;
+  height: 15rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  vertical-align: ce;
+  div {
+    height: 80%;
+    width: 80%;
+    border-style: solid;
+    border-color: white;
+    display: flex;
+    align-items: center;
+  }
+  p{
+    text-align: center;
+    font-size: map-get($fontSizes, secondary );
+    padding: 0 0.5rem;
+    color: map-get($colors, darkBrown );
+  }
+  @include sm {
+    width: 100%;
+    height: 60%;
+  }
+  @include md {
+    width: 60%;
+    height: 60%;
   }
 }
