+++ src/pages/BookListBySearch.jsx
@@ -1,10 +1,12 @@
 import React, { useEffect, useState } from "react";
+import { Helmet } from "react-helmet";
 import { useParams } from "react-router-dom";
 import BookList from "../components/Main/BookList";
 import dataTransformation from "../utils/dataTranformationBookSearch";
 import useFetch from "../utils/useFetch";
 import { useDispatch, useSelector } from "react-redux";
 import { paginationActions } from "../store/pagination";
+import optionsDropdown from "../utils/searchBarDropdownOptions";
 
 function BookListBySearch() {
   const [books, setBooks] = useState([]);
@@ -14,6 +16,8 @@ function BookListBySearch() {
 
   const offset = useSelector((state) => state.pagination.offset);
   const limit = useSelector((state) => state.pagination.limit);
+
+  const typeOfSearch = optionsDropdown.find(opt => opt.id === params.searchType).name;
 
   const convertedParams = params.searchField.replace(/\s/g, "+");
   const { data, isLoading, error } = useFetch(
@@ -32,11 +36,12 @@ function BookListBySearch() {
   }, [data]);
 
   return (
-    <BookList
-      isLoading={isLoading}
-      error={error}
-      books={books}
-    />
+    <div>
+      <Helmet>
+        <title>{typeOfSearch} de {convertedParams}</title>
+      </Helmet>
+      <BookList isLoading={isLoading} error={error} books={books} />
+    </div>
   );
 }
