+++ src/pages/BooksListBySubject.jsx
@@ -1,10 +1,12 @@
 import React, { useEffect, useState } from "react";
+import { Helmet } from "react-helmet";
 import { useParams } from "react-router-dom";
 import BookList from "../components/Main/BookList";
 import useFetch from "../utils/useFetch";
 import dataTransformation from "../utils/dataTransformationBookSubject";
 import { useDispatch, useSelector } from "react-redux";
 import { paginationActions } from "../store/pagination";
+import subjectsLists from "../utils/subjectsList";
 
 function BooksListBySubject() {
   const [books, setBooks] = useState([]);
@@ -14,6 +16,8 @@ function BooksListBySubject() {
 
   const offset = useSelector((state) => state.pagination.offset);
   const limit = useSelector((state) => state.pagination.limit);
+
+  const subject = subjectsLists.find(sub => sub.id === params.subject).name;
 
   const { data, isLoading, error } = useFetch(
     `https://openlibrary.org/subjects/${params.subject}.json?limit=${limit}&offset=${offset}`
@@ -32,11 +36,12 @@ function BooksListBySubject() {
   }, [data]);
 
   return (
-    <BookList
-      isLoading={isLoading}
-      error={error}
-      books={books}
-    />
+    <div>
+      <Helmet>
+        <title>Libros de {subject}</title>
+      </Helmet>
+      <BookList isLoading={isLoading} error={error} books={books} />
+    </div>
   );
 }
