+++ src/pages/Home.jsx
@@ -1,9 +1,15 @@
-import React from 'react'
+import React from "react";
+import { Helmet } from "react-helmet";
 
 function Home() {
   return (
-    <div>Home</div>
-  )
+    <div>
+      <Helmet>
+        <title>Home page </title>
+      </Helmet>
+      Home
+    </div>
+  );
 }
 
-export default Home
+export default Home;
