+++ src/utils/searchBarDropdownOptions.js
@@ -1,0 +1,8 @@
+const optionsDropdown = [
+  { id: "q", name: "Todos" },
+  { id: "author", name: "Autor" },
+  { id: "title", name: "Título" },
+  { id: "subject", name: "Categoría" },
+];
+
+export default optionsDropdown;
