+++ src/utils/subjectsList.js
@@ -1,0 +1,19 @@
+const subjectsLists = [
+  { name: "Fantasia", id: "fantasy" },
+  { name: "Cocina", id: "cooking" },
+  { name: "Romance", id: "romance" },
+  { name: "Amor", id: "love" },
+  { name: "Ciencia ficción", id: "science_fiction" },
+  { name: "Literatura juvenil", id: "juvenile_literature" },
+  { name: "Gatos", id: "cats" },
+  { name: "Programación", id: "programming" },
+  { name: "Suspenso", id: "thriller" },
+  { name: "Emprendimiento", id: "entrepreneurship" },
+  { name: "Humor", id: "humor" },
+  { name: "Literatura", id: "literature" },
+  { name: "Administración", id: "management" },
+  { name: "Ciencia ficción juvenil", id: "young_adult_fiction" },
+  { name: "Magia", id: "magic" },
+];
+
+export default subjectsLists;
